+++ src/qx/Circuit.cpp
@@ -1,11 +1,9 @@
 #include "qx/Circuit.hpp"
-#include "qx/Core.hpp"
 #include "qx/GateConvertor.hpp"
-#include "qx/InstructionExecutor.hpp"
 #include "qx/Instructions.hpp"
-#include "qx/Random.hpp"
+#include "qx/SimulationResult.hpp"
 
-#include <algorithm>
+#include <algorithm>  // for_each
 
 
 namespace qx {
@@ -24,55 +22,26 @@ Circuit::Circuit(V3OneProgram &program, RegisterManager &register_manager)
     return register_manager_;
 }
 
-void Circuit::add_instruction(Instruction instruction, ControlBits control_bits) {
-    controlled_instructions_.emplace_back(std::move(instruction), std::move(control_bits));
+void Circuit::add_instruction(std::shared_ptr<Instruction> instruction) {
+    instructions_.emplace_back(std::move(instruction));
 }
 
-[[nodiscard]] SimulationIterationResult Circuit::execute(error_models::ErrorModel const &errorModel) const {
-    InstructionExecutor instruction_executor{ register_manager_ };
-    auto &simulation_iteration_result = instruction_executor.getSimulationIterationResult();
-
-    for (auto const &controlled_instruction: controlled_instructions_) {
-        if (auto *depolarizing_channel = std::get_if<error_models::DepolarizingChannel>(&errorModel)) {
-            depolarizing_channel->addError(simulation_iteration_result.state);
-        } else {
-            assert(std::get_if<std::monostate>(&errorModel) && "Unimplemented error model");
-        }
-
-        auto const &controlBits = controlled_instruction.control_bits;
-        if (controlBits) {
-            auto isBitNotSet = [&simulation_iteration_result](auto const &cb) {
-                return !simulation_iteration_result.measurement_register.test(cb.value);
-            };
-            if (std::any_of(controlBits->begin(), controlBits->end(), isBitNotSet)) {
-                continue;
-            }
-        }
-
-        auto const &instruction = controlled_instruction.instruction;
-
-        // AppleClang doesn't support std::visit
-        // std::visit(instructionExecutor, instruction);
-        if (auto *measure = std::get_if<Measure>(&instruction)) {
-            instruction_executor(*measure);
-        } else if (auto *reset = std::get_if<Reset>(&instruction)) {
-            instruction_executor(*reset);
-        } else if (auto *reset_all = std::get_if<ResetAll>(&instruction)) {
-            instruction_executor(*reset_all);
-        } else if (auto *classicalOp = std::get_if<MeasurementRegisterOperation>(&instruction)) {
-            instruction_executor(*classicalOp);
-        } else if (auto *instruction1 = std::get_if<Unitary<1>>(&instruction)) {
-            instruction_executor(*instruction1);
-        } else if (auto *instruction2 = std::get_if<Unitary<2>>(&instruction)) {
-            instruction_executor(*instruction2);
-        } else if (auto *instruction3 = std::get_if<Unitary<3>>(&instruction)) {
-            instruction_executor(*instruction3);
-        } else {
-            assert(false && "Unimplemented circuit instruction");
-        }
+/* static */ void Circuit::add_error(SimulationIterationContext &context, error_models::ErrorModel const &errorModel) {
+    if (auto *depolarizing_channel = std::get_if<error_models::DepolarizingChannel>(&errorModel)) {
+        depolarizing_channel->addError(context.state);
+    } else if (!std::get_if<std::monostate>(&errorModel)){
+        throw std::runtime_error{ "Unimplemented error model" };
     }
+}
 
-    return simulation_iteration_result;
+[[nodiscard]] SimulationIterationContext Circuit::execute(error_models::ErrorModel const &errorModel) const {
+    auto context = SimulationIterationContext{ register_manager_ };
+    std::for_each(instructions_.begin(), instructions_.end(),
+        [&context, &errorModel](auto const& instruction) {
+            add_error(context, errorModel);
+            instruction->execute(context);
+    });
+    return context;
 }
 
 } // namespace qx