+++ src/qx/GateConvertor.cpp
@@ -1,10 +1,11 @@
 #include "qx/Core.hpp"
 #include "qx/GateConvertor.hpp"
 #include "qx/Gates.hpp"
-#include "qx/Instructions.hpp" // Measure, Unitary
+#include "qx/Instructions.hpp"  // Measure, Unitary
 #include "qx/OperandsHelper.hpp"
 
 #include <fmt/format.h>
+#include <memory>  // make_shared
 
 
 namespace qx {
@@ -23,7 +24,7 @@ void GateConvertor::visit_node(V3Node &) {
 
 template <std::size_t NumberOfQubitOperands>
 void GateConvertor::addGates(
-    core::DenseUnitaryMatrix<1 << NumberOfQubitOperands> matrix,
+    matrix_t<NumberOfQubitOperands> matrix,
     std::array<V3Many<V3ConstInt>, NumberOfQubitOperands> operands) {
     static_assert(NumberOfQubitOperands > 0);
 
@@ -35,16 +36,13 @@ void GateConvertor::addGates(
 #endif
 
     for (std::size_t i = 0; i < operands[0].size(); ++i) {
-        std::array<core::QubitIndex, NumberOfQubitOperands> ops{};
+        operands_t<NumberOfQubitOperands> ops{};
         for (std::size_t op = 0; op < NumberOfQubitOperands; ++op) {
             ops[op] = core::QubitIndex{
                 static_cast<std::size_t>(operands[op][i]->value)};
         }
 
-        Unitary<NumberOfQubitOperands> unitary{matrix, ops};
-
-        auto controlBits = std::make_shared<std::vector<core::QubitIndex>>();
-        circuit_.add_instruction(unitary, controlBits);
+        circuit_.add_instruction(std::make_shared<Unitary<NumberOfQubitOperands>>(matrix, ops));
     }
 }
 
@@ -144,27 +142,21 @@ void GateConvertor::addGates(const V3Instruction &instruction) {
         // The right-hand side operand, q, is the operand 1
         const auto &bit_indices = operands_helper.get_register_operand(0);
         const auto &qubit_indices = operands_helper.get_register_operand(1);
-        auto controlBits = std::make_shared<std::vector<core::QubitIndex>>();
         for (size_t i{ 0 }; i < qubit_indices.size(); ++i) {
-            circuit_.add_instruction(
-                Measure{
-                    core::QubitIndex{ static_cast<std::size_t>(qubit_indices[i]->value) },
-                    core::BitIndex{ static_cast<std::size_t>(bit_indices[i]->value) }
-                },
-                controlBits);
+            circuit_.add_instruction(std::make_shared<Measure>(
+                core::QubitIndex{ static_cast<std::size_t>(qubit_indices[i]->value) },
+                core::BitIndex{ static_cast<std::size_t>(bit_indices[i]->value) }
+            ));
         }
     } else if (name == "reset") {
-        auto controlBits = std::make_shared<std::vector<core::QubitIndex>>();
         if (instruction.operands.empty()) {
-            circuit_.add_instruction(ResetAll{}, controlBits);
+            circuit_.add_instruction(std::make_shared<Reset>(std::nullopt));
         } else {
             const auto &qubit_indices = operands_helper.get_register_operand(0);
             for (size_t i{ 0 }; i < qubit_indices.size(); ++i) {
-                circuit_.add_instruction(
-                    Reset{
-                        core::QubitIndex{ static_cast<std::size_t>(qubit_indices[i]->value) }
-                    },
-                    controlBits);
+                circuit_.add_instruction(std::make_shared<Reset>(
+                    core::QubitIndex{ static_cast<std::size_t>(qubit_indices[i]->value) }
+                ));
             }
         }
     } else {