+++ CHANGELOG.md
@@ -11,7 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Take measurement register out of `QuantumState`. 
-- Change `Circuit::execute` to return a `SimulationIterationResult`.
+- Change `Circuit::execute` to return a `SimulationIterationContext`.
+- Implement instructions as a hierarchy. 
 
 ### Removed